Refactor: move badges to ui/components dir
