+++ docs/PRD.md
@@ -427,8 +427,6 @@ src/
 ├── todoist.rs                 # API models & display structs
 ├── sync.rs                    # Sync service
 ├── storage.rs                 # SQLite storage (in-memory)
-├── badge.rs                   # Badge creation utilities
-├── terminal_badge.rs          # Terminal-optimized badges
 ├── icons.rs                   # Icon service
 └── ui/                        # TUI components
     ├── app.rs                 # Application state management
@@ -436,6 +434,7 @@ src/
     ├── layout.rs              # Layout management
     ├── renderer.rs            # Main rendering loop
     └── components/            # UI components
+        ├── badge.rs           # Terminal-optimized badges
         ├── projects_list.rs   # Projects display
         ├── tasks_list.rs      # Tasks display
         ├── status_bar.rs      # Status bar
