+++ src/main.rs
@@ -1,9 +1,7 @@
-pub mod badge;
 pub mod debug_logger;
 pub mod icons;
 pub mod storage;
 pub mod sync;
-pub mod terminal_badge;
 pub mod todoist;
 pub mod ui;
