+++ src/ui/components/badge.rs
@@ -7,7 +7,7 @@ use crate::todoist::LabelDisplay;
 
 #[derive(Debug, Clone, Copy)]
 /// Badge styles optimized for terminal compatibility
-pub enum TerminalBadgeStyle {
+pub enum BadgeStyle {
     Primary,
     Success,
     Warning,
@@ -19,38 +19,38 @@ pub enum TerminalBadgeStyle {
     Underlined,
 }
 
-impl TerminalBadgeStyle {
+impl BadgeStyle {
     fn to_style(self) -> Style {
         match self {
             // Use bright colors with proper backgrounds for better visibility
-            TerminalBadgeStyle::Primary => Style::default()
+            BadgeStyle::Primary => Style::default()
                 .fg(Color::Black)
                 .bg(Color::LightBlue)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Success => Style::default()
+            BadgeStyle::Success => Style::default()
                 .fg(Color::Black)
                 .bg(Color::LightGreen)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Warning => Style::default()
+            BadgeStyle::Warning => Style::default()
                 .fg(Color::Black)
                 .bg(Color::LightYellow)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Danger => Style::default()
+            BadgeStyle::Danger => Style::default()
                 .fg(Color::White)
                 .bg(Color::Red)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Info => Style::default()
+            BadgeStyle::Info => Style::default()
                 .fg(Color::Black)
                 .bg(Color::LightCyan)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Secondary => Style::default()
+            BadgeStyle::Secondary => Style::default()
                 .fg(Color::White)
                 .bg(Color::Gray)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Bordered | TerminalBadgeStyle::Bold => Style::default()
+            BadgeStyle::Bordered | BadgeStyle::Bold => Style::default()
                 .fg(Color::White)
                 .add_modifier(Modifier::BOLD),
-            TerminalBadgeStyle::Underlined => Style::default()
+            BadgeStyle::Underlined => Style::default()
                 .fg(Color::White)
                 .add_modifier(Modifier::UNDERLINED | Modifier::BOLD),
         }
@@ -59,25 +59,25 @@ impl TerminalBadgeStyle {
 
 /// Create a terminal-optimized badge with text and style
 #[must_use]
-pub fn create_terminal_badge(text: &str, style: TerminalBadgeStyle) -> Span<'static> {
+pub fn create_badge(text: &str, style: BadgeStyle) -> Span<'static> {
     Span::styled(format!(" {text} "), style.to_style())
 }
 
 /// Create badges with brackets (ASCII fallback)
 #[must_use]
-pub fn create_bracket_badge(text: &str, style: TerminalBadgeStyle) -> Span<'static> {
+pub fn create_bracket_badge(text: &str, style: BadgeStyle) -> Span<'static> {
     Span::styled(format!("[{text}]"), style.to_style())
 }
 
 /// Create badges with parentheses
 #[must_use]
-pub fn create_paren_badge(text: &str, style: TerminalBadgeStyle) -> Span<'static> {
+pub fn create_paren_badge(text: &str, style: BadgeStyle) -> Span<'static> {
     Span::styled(format!("({text})"), style.to_style())
 }
 
 /// Create a label badge with custom color
 #[must_use]
-pub fn create_terminal_label_badge(name: &str, color: &str) -> Span<'static> {
+pub fn create_label_badge(name: &str, color: &str) -> Span<'static> {
     // Convert Todoist color names to terminal colors
     let bg_color = match color.to_lowercase().as_str() {
         "red" => Color::Red,
@@ -106,7 +106,7 @@ pub fn create_terminal_label_badge(name: &str, color: &str) -> Span<'static> {
 
 /// Create task badges optimized for terminal compatibility
 #[must_use]
-pub fn create_terminal_task_badges(
+pub fn create_task_badges(
     is_recurring: bool,
     has_deadline: bool,
     duration: Option<&str>,
@@ -115,20 +115,20 @@ pub fn create_terminal_task_badges(
     let mut badges = Vec::new();
 
     if is_recurring {
-        badges.push(create_bracket_badge("REC", TerminalBadgeStyle::Primary));
+        badges.push(create_bracket_badge("REC", BadgeStyle::Primary));
     }
 
     if has_deadline {
-        badges.push(create_bracket_badge("DUE", TerminalBadgeStyle::Danger));
+        badges.push(create_bracket_badge("DUE", BadgeStyle::Danger));
     }
 
     if let Some(duration) = duration {
-        badges.push(create_paren_badge(duration, TerminalBadgeStyle::Warning));
+        badges.push(create_paren_badge(duration, BadgeStyle::Warning));
     }
 
     // Add label badges
     for label in labels {
-        badges.push(create_terminal_label_badge(&label.name, &label.color));
+        badges.push(create_label_badge(&label.name, &label.color));
     }
 
     badges
@@ -136,12 +136,12 @@ pub fn create_terminal_task_badges(
 
 /// Create priority badges with better terminal support
 #[must_use]
-pub fn create_terminal_priority_badge(priority: i32) -> Option<Span<'static>> {
+pub fn create_priority_badge(priority: i32) -> Option<Span<'static>> {
     match priority {
-        4 => Some(create_terminal_badge("P0", TerminalBadgeStyle::Danger)), // Urgent
-        3 => Some(create_terminal_badge("P1", TerminalBadgeStyle::Warning)), // High
-        2 => Some(create_terminal_badge("P2", TerminalBadgeStyle::Info)),   // Medium
-        1 => Some(create_terminal_badge("P3", TerminalBadgeStyle::Secondary)), // Low
-        _ => None,                                                          // No priority
+        4 => Some(create_badge("P0", BadgeStyle::Danger)),    // Urgent
+        3 => Some(create_badge("P1", BadgeStyle::Warning)),   // High
+        2 => Some(create_badge("P2", BadgeStyle::Info)),      // Medium
+        1 => Some(create_badge("P3", BadgeStyle::Secondary)), // Low
+        _ => None,                                            // No priority
     }
 }
